S328-046 Introduce asynchronous processing
Introduce a separate task for the actual processing, synchronized
to the main task via a queue of requests. This allows the main
task to never be blocking on input when libadalang is performing
an expensive computation. This will also be used to implement
a basic support for cancelling requests.

Also process the output in a separate task.