S603-029  implement "is called by" as a custom request
Introduce a custom request "textDocument/ALS_called_by".

Add a subprogram in LSP.LAL_Utils to implement the functionality,
add support for marshalling/unmarshalling the response to the
JSON stream, and add support for serving this request in the LSP
server.

Move the utility function Find_All_References to LSP.LAL_Utils.

Add a test for the 'called_by' functionality.