S624-027 Add adaptation for the new API in Ada.Directories
The behavior of Ada.Directories.Containing_Directory has
changed slightly. The code that added support for this was
not taking into account the situation on Windows; this is
now done.

Clarify the corresponding comment to say that this is an
API change in Ada.Directories, not in GNAT per se.