S815-010 S709-013 Initial code for multi-project support
This is a first step for supporting multiple projects: add support
for more than one Context, and organise the requests so that
they either get the results from all contexts and aggregate
them, or only query the results from the appropriate context.

At the moment, there is only one project loaded, and we have
added one context to hold all files that are not associated
to a project.

Move the diagnostics preference from the Context to the
Message_Handler.