S910-023 Add support for aggregate projects
Add support for aggregate projects. To do this, move
project loading from the Context to the Message_Handler: the
Message_Handler is responsible for creating one Context per
project hierarchy within the aggregate project.

Add a first test to check the result of finding all references
in a simple aggregate project tree.