SC06-075 Separate Documents from Contexts
Documents should not belong to specific contexts. For instance, a
document might represent an Ada source which is common to several
project contexts. Following this, a document should not own its
Analysis_Unit, the Analysis_Unit is valid only for a specific context.

Refactor the code in light of this: the documents are now held by
the Message_Handler.

Add a test to check that references for an entity that's common to several
project works after editing the document.