SC20-022 'implementation' now returns bodies for parent/child methods
Similarly to 'declaration', this requests now lists the implementations
of parent and child definitions of the subprogram.

Also slightly improve the behavior of 'implementation': return the
declaration if it's an "is null" procedure or an expression function.

Fix a bug that returned no results for subprograms that didn't have
a declaration.

Add a test.