SB25-033 Remove performance bottleneck
When returning a list of locations, we were checking at each insertion
whether the location was already present in the results. This has
a cost of O(n^2)

Instead, append the locations, and sort/remove duplicates only at
the end, changing the cost to O(n*log(n)).

Sort the results by file then location; when sorting files in the
same directory, use a slightly more logical sort than an alphabetic
sort, returning locations in the .ads before the .adb, and the parent
package before the child packages.