T123-038 Improve project handling, and clean code up
In the case that no project is specified, use an implicit
project which covers the subdirectories as well. This is slightly
friendlier for VScode users who don't have a .gpr at the root
of a hierarchy. Adapt test baselines in light of this.

Now that documents are owned by the handler, there is no longer
the need for a "projectless" context. Clean up the code in light
of this.