T123-048 Implement support for receiving incremental editor changes
For now the data structure implementing the in-memory contents
of the editor is an Unbounded_Wide_String. To speed up the lookup
of locations given in line,character coordinates, add an index
containing the line numbers.

For now, this is controlled via a trace which is deactivated by default,
until we get more field testing.

Add tests for some edge cases.