T306-021: Add memory usage monitor based on GNATCOLL.Memory
when the DEBUG.ADA_MEMORY trace is enabled, we now monitor the
memory usage consumption and dump it in the logs.

This allows to identify which parts of the program are the
biggest memory usage consumers.

The stacktrace addresses are resolved using GNAT.Symbolic.Traceback
because we link the ALS with relocatable libraries.