T319-027: Handle package prefixes for subp calls completion
We now use the new libadalang Is_Dot_Call function on the returned
compeltion items to know if we are dealing with a dotted subporgram call
or not, allowing us to distiguish dotted calls, for which we should omit
the first parameter in the returned completion snippets, from subprogram
calls that are prefixed by a package names (e.g: Ada.Text_IO.Put_Line).

An automatic test has also been added.