SA14-008: Handle aggregates for completion
Completion snippets that list all the possible shapes (discriminants,
components...) of an aggregate are now proposed when opening a parenthesis
to initialize/assign records.

A new flag has also been added in order to control the number of parameters
at which point named notation should be used in the proposed snippets.

Tests have been adapted since "(" now can trigger completion and thus is
listed in the LSP triggerCharacters config field.