T318-064: Provide completion for predefined item
We now provide basic completion for keywords, aspects, pragmas and
attributes.

While keyword completion is provided directly by LAL, completion items
for pragmas, aspects and attributes are read through a JSON database.

Currently we only have a database for Ada 2012 aspects, pragmas and
keywords: the plan is to have one database per Ada version, and
determine which database should be used depending on the GNAT install
found in the PATH.

Thus, this is work in progress: we need to add a database per version
and we need to refine the behavior for keyword and attributes'
completion (some LAL bits are missing).