T122-009: Add a 'displayMethodAncestryOnNavigation' configuration flag
This flag allows to customize the behavior of the navigation requests
(e.g: textDocument/definition) regarding the listing of overriding/overridden
subprograms.

A dedicated setting has been added to the VS Code extension.

We don't list overridding/overridden subprograms on textDocument/declaration
anymore, no matter the value of the preference: if we are on a body, we should
always go to the corresponding declaration. The user can then navigate through
the hierarchy from the declaration.

Tests have been adjusted: more tests will come once we fully agree on the
behavior.