T812-022: Imprecise results for text/Definition
Better handling to navigate between the declaration/body/ref
when the prototype are slightly differents (the common case
is when the user is modifying the declaration and then when to
do the same with the body)

Add a test.
Adapt existing tests to the new results found by the Imprecise
Fallback.