TA09-019 Add support for filesystem monitoring
This adds support for monitoring the filesystem for changes
that might occur in project directories, outside of the IDE.

Add support for `workspace/didChangeWatchedFiles`.

Add an optional dependency on ada_libfswatch.gpr, and document
it. Add support in the travis CI builder to generate it on Linux.
Add a stub project to make the ALS useable without this dependency.

Rework the way files are indexed, to allow reindexing for any
file that might change on disk. In particular, no longer store
in LSP.Ada_Contexts a reference to the "Last_Indexed" file, which
prevented indexing of new files altogether.