T127-013 Fix support for encoding Code Action commands
The ALS uses a system based on the Ada Tags for encoding/
decoding commands associated to a code action. This requires
types with specific tags to be declared: this is the case
for the ALS server but should not be required for LSP clients.

Add defensive programming to allow this.