U308-004 U305-040 Minor update on exception handling
Catch Predicondition_Failure in the same way we intercept
Property_Error, ie assuming that this can happen in the normal
flow when working on invalid code, and not report them to the
IDE.

Systematically store the symbolic backtrace of exceptions in the
log.

Factorize the code to do this for both notifications and requests.