U304-005 Add basic support for preprocessor directives
For now, we do not support an actual preprocessor, but work
in a degraded mode where the "#if" branch of preprocessing is
always chosen.

This allows responding to requests in cases where simply
"ignoring all preprocessing" is not sufficient because it
results in non-valid syntax, for instance

     #if A
        procedure foo (
     # else
        procedure foo (
     # endif

We do this by intercepting buffers before they are sent
to Libadalang, and preprocess on-the-fly.

This also adds support for working with non-LF line terminators:
the line terminators are changed to LF as part of the
preprocessing.