U415-038: Change implementation of callHierarchy requests
We now use the same code that we use for our old and custom
alsCalledBy and alsCalls requests.

The next step is to remove our custom requests and use only
the standard callHierarchy LSP requests instead.

The associated tests have been removed: we'll reuse the tests
that we have created for our custom requests instead, which
cover way more cases.