U301-023 Introduce LSP_URI type
to ensure that all URI are normalized on reading from JSON.
We need this to avoid distinct names for the same file
on Windows (e.g. `z:/file` and `Z:/file`).

The idea is to normalize all URI read from JSON and all
file names got from GNATCOLL.Projects. We keep all files
from Libadalang unnormalized in hope that libadalang won't
change file names. So we normalize file in the libadalang
unit provider.