U805-023: use previous context in signatureHelp
If available try to reuse the signatureHelpContext which contains
the previous signatures: it's faster to just filter them that
recompute a new list of signatures.

Edge case for "\b": removing a character can make more signatures
valid and they could not be contained in the previous signature =>
always recompute the list.

Adapt tests: retriggerCharacters have changed
Add a new test.