U729-020 Don't send invisible name in completion if
there is some visible name with the same label in result.

For instance, if we have `Pkg.Ada` in the project and
`with Ada;` in unit clauses, we shouldn't propose `Pkg.Ada`
as invisible completion for `Ada` identifier, because it
won't work any way. `Ada` is already visible and the
second package can't become visible even if we append
`with Pkg.Ada; use Pkg;` clauses.