U622-021: Compute some completion items properties lazily
Use the newly added completionItem/resolve request to compute
the 'documentation' and 'detail' fieds of completion items lazily
when the client supports it (i.e: when a given item gets selected
in the completion window).

This makes the ALS way faster to return completion items for packages
that contain a huge number of declarations (e.g: Libadalang.Common).

VS Code already supports this lazy computation method but not other
clients.