UA26-012: Don't set isIncomplete to True for completion
Since we don't properly handle it in GS and we don't limit
the results we write in the ALS side.

This has just for effect to slow down completion on VS Code, since
it tries to retrigger it on each character while we already return
all the results the first time.

Adapt test to make it more robust.