V303-003 Make sure to call Get_From_Buffer only on Ada files
Calling Libadalang.Analysis.Get_From_Buffer on non-Ada files
might claim a lot of memory, as LAL will attempt to read it
as an Ada file and potentially produce a large amount of
diagnostics.

To avoid this, centralize all the calls to Get_From_Buffer
in a LSP.Ada_Contexts.Get_AU, which filters out non-Ada files.
(This also makes sure that we're passing the proper charset
to Get_From_Buffer).