UB25-027: Add a custom request to list source directories
Allowing the extension to query them in order to add them to the workspace
if not already present.

This allows to search for files that come from imported projects that might
not be under the workspace root directory.

This works only for importing non-externally built source directories
for the moment.