V705-026: Compute predefined types' completion doc immediately
When computing completion items' details, compute the documentation
immediately for predefined types (e.g: Integer) since the standard
package does not correspond to any actual URI.

This fixes an exception on the completionItem/resolve request, where
we are trying to retrieve a node from an URI and a SLOC which
does not actually exist.

Add an automatic test.