V511-027: Run PP on completion snippet
Encode snippet in pseudo Ada code, run PP and decode it.
When successfull, it will generate a textEdit for the
completionItem.

Enforce PP switches to have the parameter aligned: it can
be disabled via a trace which is used to avoid updating
all the snippet tests.

Add tests with the trace disabled to check the PP behavior
on snippets.