Add support for setTrace notification
In order to not raise an exception when VS Code sends it.

We don't handle it for now since VS Code already outputs a lot
of useful information (resquests being sent, results...) when
the trace setting is set to 'messages' or 'verbose'.

This could be used in the future if we want to have more debug
traces in the ALS.

closes eng/ide/ada_language_server#1116