Emit a diagnostic if no runtime found.
After loading a project, check if the runtime was found.
In case when we found no runtime change the status of loaded
project accordingly. Use this status to emit the diagnostic
on each opened file to let the user know that runtime was
not found.

Report GPR2 warnings in (in addition to errors) to the client,
because if GPR2 can't find runtime then it emits a warning.
Adapt tests with dummy runtime settings.
Fix #1128