Clarify the message in hovers when in aggregate projects
When an aggregate project is loaded, in the interest of performance,
we don't query all contexts when resolving a tooltip, but only resolve
for the first project in which the file belong.

In this case, we annotate the tooltip text with a mention of the
project. This change clarifies this mention so that it cannot be interpreted
as a failure to resolve requests in the rest of the toplevel projects.

A test existed to exercise this: update the baseline to match the
text that we now expect.

For W505-020