Restore als-source-dirs command for VS Code popup
In order to create a multi-root workspace when some project's
source directories are not located under the VS Code workspace's
root directory, which may happen  when the loaded porject imports
other projects.

Checks have also been added in order to not display the popup when
the user has already specified some folders explcitly in his settings
or if the folder to add is already present.

For eng/ide/ada_language_server#1190