Improve auto import suggestion
When the name is already qualified, use the term "Import from ", which
implies that a with clause will be added but the qualifier won't change.
When the name is not qualified or can be imported from other package
other than its qualifier, use the term "Qualify with ", which implies
that a with clause may be added if needed, and the qualifier will also
be added or updated.