Add with-clause for invisible completion items
We now append the right 'als-auto-import' command to
invisible completion items, when the 'insertWithClauses' setting
is enabled.

This allows client that support executing completionItem commands
to execute the attached command in order to add the missing with-clause,
appending the qualifier if not already specified.

Automatic tests for both cases (i.e: with and without an already specified
qualifier) have been added.

For eng/ide/ada_language_server#1159