Do not add qualifier when completing invisible units
We should only add a qualifier if the user is completing
an invisible declaration contained in a invisible unit
(e.g: Ada.Text_IO.Put_Line) not when completing the invisible
unit itself (e.g: Ada.Text_IO).

For that, check that the basic declaration attached to the completion
proposal is the same as the enclosing unit's root declaration: if yes,
we should not add any qualifier.

Add a new test case for that.

For eng/ide/ada_language_server#1159