Rewrite project status data
Create a specific package to handle project status which is a
combination of project type and a load status.

Only use implicit project for specific cases and remove it
as a fallback mechanism. To replace it, create an empty
context when the project can't be loaded.

Add a test

For eng/ide/ada_language_server#1358