Always use LAL's P_Complete for dotted name completion
Since it already handles perfectly invisible completion when
a dotted name is specified (e.g: 'Ada.Text_IO.', without a with-clause
on this unit).

In combination, we now disable the invisible completion provider on
dotted names and after an open parenthesis, to not interfer respectively
with the LAL completion provider and the parameter/aggregate completion
provider.

Adapt the tests and add a new one for child unit invisible completion
filtering.

For eng/ide/ada_language_server#1253