Fix als-other-file command for child packages
For this command, we were  calling Unit which expects an unit name
but we were passing a file base name (without extension).

So this doesn't work if the file base name contains '-'
(child packages).

Use the proper GPR2 API to get a proper unit name instead.

Add an automatic test.

For eng/ide/ada_language_server#1450