python test driver: do not spawn a separate Python
... But load directly test.py as a module, and execute the
functions marked with a given decorator.

This means we cannot change the directory for each test
(because of the possibility of running tests in parallel), so
pass the working directory to each test.

Remove initialization/initialize/didChangeConfiguration/shutdown
boilerplate from simple tests.

As a proof of concept and to test parallelism, convert another
test to the Python driver.