python test driver: on Windows, spawn a separate Python
This is needed so we can properly remove the test dir
once the test is done, without risking that the Python
interpreter has a handle on files there.

Slightly improve the semantics for the wait factor.
Improve error reporting slightly.
Fix creation of URIs with respect to drive letters.

Follow-up on #1462