Use fallback context for non-source Ada files
Allowing to always have a valid context for any opened
Ada file: either the file is known by a project context and
we use it, either the file does not belong to any project
subtree and we use the fallback context.

Having a fallback context allows to at least create a proper
LAL analysis unit for this file, which in turns allow most
of the LSP requests to work properly.

For eng/ide/ada_language_server#1517