Call Update_Sources only when GPR files are saved
Update_Sources is too slow to be called each time the
user edits the file: call it only when opening the file for
the first time and when saving it.

Add a test that checks that source files diagnostics get properly
updated when saving edited GPR files.

For eng/ide/ada_language_server#1576