Fix the navigation / find all refs in extending projects
The function GPR2.Project.View.Closure only navigates to the views
that are "with'ed" by a view - this is not satisfactory in the case
of extending projects, where we want to cover the entire extended
hierarchy even if there isn't a "with" relationship between the
extending project and the extended ones.

We fix this by providing our own function to walk a project closure.

Add a test.

Add an entry in CHANGELOG.md.

Add support for using the Python driver to conveniently make
textDocument/references requests.

Add a bit of typing information in the Python driver.

Fixes #1583.