Monitor the filesystem for changes in all source dirs
... Rather than only monitorint the source dirs for projects that are
not externally built.

This will provide a slightly better experience, as it will allow
the language server to react to the new installation of dependencies
for the working project.

This is also an opportunity for code simplification, as we were
maintaining a separate list of source directories just for this purpose.

Implements #1584.