Initialize file versions from didOpen argument
Before this patch, the language server initialized versions of all newly
opened file to 1. However, the language server protocol defines a
parameter for didOpen that says that the client gets to decide the
initial versions of opened files. One client, namely Helix 25.1, passes
an initial value of 0 which caused e.g. renaming a variable right after
opening a file to fail.

This patch makes the initial value of versions the one that's passed by
the client.