Add status bar displaying project status + commands
A status bar item is now displayed on the right side of the VS Code
status bar, showing the project-loading status and a few useful commands
when the user hovers on it.

This is a first version to gather more feedback on what we want to
display there.

For eng/ide/ada_language_server#1520