Add a new diagnostics' source for files
This diagnostic source emits hint diagnostics for files opened
that do not belong to the loaded project tree (or that no
project tree was loaded), warning the user
that some requests might not work as expected.

A VS Code setting has been added for this new diagnostics' source.

Add tests for this.

For eng/ide/ada_language_server#1593