Remove .gnatdebug support in the ALS
In practice we don't use it, because we already have the
--tracefile option.

And it can be dangerous if a .gnatdebug file targetting other
software (e.g: GS or GPRbuild) is present in the ALS working
directory.

Add a test that verifies that the server does not crash when
a .gnatdebug file is present in the working directory.

For eng/ide/ada_language_server#1663