Make problemMatchers test more robust
By omitting the last part of GNAT messages that lists the switch
used to detect the message (e.g: [-gnatwm]).

For an unknown reason, this part gets truncated on Windows somehow
for one specific message, so let's just skip it: we just want to
check that the severities are correctly set for each message.

For eng/ide/ada_language_server#1696