Add labelDetails to completion switches
If the client supports it, which is the case of VS Code.

labelDetails are displayed in gray, next to the completion
item's label: displaying the doc here gives a preview to
the user, which is particularly useful when trying to
understand what the selected switch is doing.

Add a test.

For eng/ide/ada_language_server#1741